+++ alembic.ini
@@ -38,8 +38,7 @@ prepend_sys_path = .
 # the output encoding used when revision files
 # are written from script.py.mako
 # output_encoding = utf-8
-
-sqlalchemy.url = driver://user:pass@localhost/dbname
+sqlalchemy.url = mysql://zaim:zaim@localhost:3306/oumpsy
 
 
 [post_write_hooks]
