+++ alembic/versions/604740cb89b9_added_id_colum.py
@@ -1,0 +1,28 @@
+"""Added id colum
+
+Revision ID: 604740cb89b9
+Revises: 0248b0da327b
+Create Date: 2022-10-08 13:07:36.640063
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '604740cb89b9'
+down_revision = '0248b0da327b'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('psy_code', sa.Column('id', sa.String(length=50), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('psy_code', 'id')
+    # ### end Alembic commands ###
