+++ tajma/model/RoleModel.py
@@ -1,0 +1,7 @@
+from tajma import Base
+from sqlalchemy import Column, String
+
+class Role(Base):
+    __tablename__="psy_role"
+    id = Column(String(50), primary_key=True)
+    name = Column(String(50), unique=True)
