+++ tajma/verify.py
@@ -2,7 +2,8 @@ from flask import flash
 from tajma import bcrypt
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField
-from tajma.models import Student, User
+# from tajma.models import Student, User
+from tajma.model.StudentModel import Student
 from wtforms.validators import DataRequired, Email
 
 class VerificationForm(FlaskForm):
@@ -15,7 +16,7 @@ class VerificationForm(FlaskForm):
             #check if email inside Student table
             if self.email.data == u.email:
                 flash("We found your email")
-             return True
+            return True
         flash("Email not found")   
         return False
